refactor(dashboard-api): flatten wrappers, fold sso handler tests into team tests
- identity is now a required value param of bootstrapUser (production always
  has one since the OIDC-only path); the nil-identity guards, the
  setOIDCIdentityExternalID wrapper, and the test-only nil-identity
  concurrency test go away — s.idp.SetIdentityExternalID is called directly
- inline the one-expression creatorContextFromSignupMetadata
- drop zero-value test constructors (newTestIdentityProvider et al)
- rename provisioning_error.go -> utils_provisioning.go per handlers convention
- move the two PostTeamsTeamIDMembers SSO-guard tests from sso_test.go into
  team_handlers_test.go where the other member tests live
