+++ packages/dashboard-api/internal/provisioning/bootstrap.go
@@ -24,10 +24,14 @@ func (s *Service) BootstrapOIDCUser(ctx context.Context, input OIDCUserBootstrap
 		UserID:          uuid.New(),
 		Email:           input.OIDCUserEmail,
 		DefaultTeamName: defaultTeamNameFromOIDCUserName(input.OIDCUserName),
-		CreatorContext:  creatorContextFromSignupMetadata(input.SignupIP, input.SignupUserAgent, teamprovision.AuthMethodSocial),
+		CreatorContext: normalizeCreatorContext(&teamprovision.CreatorContextV1{
+			IPAddress:  input.SignupIP,
+			UserAgent:  input.SignupUserAgent,
+			AuthMethod: teamprovision.AuthMethodSocial,
+		}),
 	}
 
-	return s.bootstrapUserWithIdentity(ctx, profile, &bootstrapUserIdentity{
+	return s.bootstrapUser(ctx, profile, bootstrapUserIdentity{
 		Issuer:  input.OIDCIssuer,
 		Subject: input.OIDCUserID,
 	})
@@ -48,16 +52,12 @@ func (s *Service) requireConfiguredOIDCIssuer(issuer string) error {
 	}
 }
 
-func (s *Service) bootstrapUserWithIdentity(ctx context.Context, profile bootstrapUserProfile, identity *bootstrapUserIdentity) (ProvisionedTeam, error) {
+func (s *Service) bootstrapUser(ctx context.Context, profile bootstrapUserProfile, identity bootstrapUserIdentity) (ProvisionedTeam, error) {
 	// Resolve the identity's SSO organization before opening the transaction: the
 	// Kratos lookup is a network call that must not run under the per-user lock.
-	var ssoOrgID uuid.UUID
-	if identity != nil {
-		orgID, err := s.idp.GetIdentityOrganizationID(ctx, identity.Subject)
-		if err != nil {
-			return ProvisionedTeam{}, fmt.Errorf("resolve sso organization: %w", err)
-		}
-		ssoOrgID = orgID
+	ssoOrgID, err := s.idp.GetIdentityOrganizationID(ctx, identity.Subject)
+	if err != nil {
+		return ProvisionedTeam{}, fmt.Errorf("resolve sso organization: %w", err)
 	}
 
 	authTxDB, tx, err := s.authDB.WithTx(ctx)
@@ -68,39 +68,35 @@ func (s *Service) bootstrapUserWithIdentity(ctx context.Context, profile bootstr
 		_ = tx.Rollback(ctx)
 	}()
 
-	if identity != nil {
-		existing, err := authTxDB.GetUserIdentity(ctx, authqueries.GetUserIdentityParams{
-			OidcIss: identity.Issuer,
-			OidcSub: identity.Subject,
-		})
-		switch {
-		case err == nil:
-			profile.UserID = existing.UserID
-		case !dberrors.IsNotFoundError(err):
-			return ProvisionedTeam{}, fmt.Errorf("get user identity: %w", err)
-		}
+	existing, err := authTxDB.GetUserIdentity(ctx, authqueries.GetUserIdentityParams{
+		OidcIss: identity.Issuer,
+		OidcSub: identity.Subject,
+	})
+	switch {
+	case err == nil:
+		profile.UserID = existing.UserID
+	case !dberrors.IsNotFoundError(err):
+		return ProvisionedTeam{}, fmt.Errorf("get user identity: %w", err)
 	}
 
 	candidateUserID := profile.UserID
 	if err := authTxDB.UpsertPublicUser(ctx, candidateUserID); err != nil {
 		return ProvisionedTeam{}, fmt.Errorf("upsert public user: %w", err)
 	}
-	if identity != nil {
-		canonicalUserID, err := authTxDB.UpsertPublicIdentity(ctx, authqueries.UpsertPublicIdentityParams{
-			OidcIss: identity.Issuer,
-			OidcSub: identity.Subject,
-			UserID:  candidateUserID,
-		})
-		if err != nil {
-			return ProvisionedTeam{}, fmt.Errorf("upsert public identity: %w", err)
+	canonicalUserID, err := authTxDB.UpsertPublicIdentity(ctx, authqueries.UpsertPublicIdentityParams{
+		OidcIss: identity.Issuer,
+		OidcSub: identity.Subject,
+		UserID:  candidateUserID,
+	})
+	if err != nil {
+		return ProvisionedTeam{}, fmt.Errorf("upsert public identity: %w", err)
+	}
+	if canonicalUserID != candidateUserID {
+		// concurrent bootstrap claimed the identity first; drop the orphan candidate row
+		if err := authTxDB.DeletePublicUser(ctx, candidateUserID); err != nil {
+			return ProvisionedTeam{}, fmt.Errorf("delete orphan public user: %w", err)
 		}
-		if canonicalUserID != candidateUserID {
-			// concurrent bootstrap claimed the identity first; drop the orphan candidate row
-			if err := authTxDB.DeletePublicUser(ctx, candidateUserID); err != nil {
-				return ProvisionedTeam{}, fmt.Errorf("delete orphan public user: %w", err)
-			}
-			profile.UserID = canonicalUserID
-		}
+		profile.UserID = canonicalUserID
 	}
 
 	// Serialize bootstrap for a user even when they have no team memberships yet.
@@ -130,9 +126,9 @@ func (s *Service) bootstrapUserWithIdentity(ctx context.Context, profile bootstr
 		// are durably committed, so a failed PATCH never outlives a rolled-back
 		// transaction. This is also the recovery path: a user whose external_id was
 		// never set (e.g. a prior bootstrap whose PATCH failed after commit) re-runs
-		// here and the PATCH is re-asserted. setOIDCIdentityExternalID is idempotent.
-		if err := s.setOIDCIdentityExternalID(ctx, identity, profile.UserID); err != nil {
-			return ProvisionedTeam{}, err
+		// here and the idempotent PATCH is re-asserted.
+		if err := s.idp.SetIdentityExternalID(ctx, identity.Subject, profile.UserID); err != nil {
+			return ProvisionedTeam{}, fmt.Errorf("set ory identity external id: %w", err)
 		}
 
 		return ProvisionedTeam{
@@ -160,8 +156,8 @@ func (s *Service) bootstrapUserWithIdentity(ctx context.Context, profile bootstr
 		}
 
 		// No billing: SSO teams are provisioned out of band.
-		if err := s.setOIDCIdentityExternalID(ctx, identity, profile.UserID); err != nil {
-			return ProvisionedTeam{}, err
+		if err := s.idp.SetIdentityExternalID(ctx, identity.Subject, profile.UserID); err != nil {
+			return ProvisionedTeam{}, fmt.Errorf("set ory identity external id: %w", err)
 		}
 
 		return landing, nil
@@ -192,8 +188,8 @@ func (s *Service) bootstrapUserWithIdentity(ctx context.Context, profile bootstr
 	}
 
 	// Emit the billing event before the external_id backfill: ProvisionTeam is
-	// fire-and-forget, but setOIDCIdentityExternalID returns early on failure, and
-	// the existing-team recovery path only re-fires ProvisionTeam within
+	// fire-and-forget, but the backfill returns early on failure, and the
+	// existing-team recovery path only re-fires ProvisionTeam within
 	// bootstrapProvisionRetryAge of team creation. Provisioning first guarantees the
 	// freshly committed team is never left billing-orphaned by an Ory PATCH failure.
 	req := teamprovision.TeamBillingProvisionRequestedV1{
@@ -210,8 +206,8 @@ func (s *Service) bootstrapUserWithIdentity(ctx context.Context, profile bootstr
 	// A PATCH failure here is recoverable: external_id stays unset, the dashboard
 	// re-runs bootstrap on the next login and re-asserts it via the existing-team
 	// path above.
-	if err := s.setOIDCIdentityExternalID(ctx, identity, profile.UserID); err != nil {
-		return ProvisionedTeam{}, err
+	if err := s.idp.SetIdentityExternalID(ctx, identity.Subject, profile.UserID); err != nil {
+		return ProvisionedTeam{}, fmt.Errorf("set ory identity external id: %w", err)
 	}
 
 	return ProvisionedTeam{
@@ -222,18 +218,4 @@ func (s *Service) bootstrapUserWithIdentity(ctx context.Context, profile bootstr
 		IsBlocked:     team.IsBlocked,
 		BlockedReason: team.BlockedReason,
 	}, nil
-}
-
-// setOIDCIdentityExternalID stores the canonical public.users id on the Ory
-// identity. It is a no-op for non-OIDC bootstrap (identity == nil).
-func (s *Service) setOIDCIdentityExternalID(ctx context.Context, identity *bootstrapUserIdentity, userID uuid.UUID) error {
-	if identity == nil {
-		return nil
-	}
-
-	if err := s.idp.SetIdentityExternalID(ctx, identity.Subject, userID); err != nil {
-		return fmt.Errorf("set ory identity external id: %w", err)
-	}
-
-	return nil
 }
